Minor improvement for TLS1.3 PSK server callback
The TLS1.3 PSK server callback has to set the ciphersuite for the PSK it
hands back to WolfSSL. This is no problem for session tickets, as they
store their related ciphersuite.
However, for external callbacks, that necessity causes problems: When a
single PSK should be used for multiple cipersuites (as long as the Hash
algorithm for the binder matches), one does not safely know which
ciphersuite to provide in the callback, as you have to provide the
currently selected one from the server during parsing of the list
provided from the client in the ClientHello.

With this improvement, we pre-set the ciphersuite pointer to the
currently selected ciphersuite from the server. The user can still
override that with another ciphersuite (no API change). However, now the
user can just leave that ciphersuite as is in case the given PSK and its
identity suites that ciphersuite.

Signed-off-by: Tobias Frauenschläger <tobias.frauenschlaeger@oth-regensburg.de>