Enable support for mandatory PSKs
Add a new option, WOLFSSL_VERIFY_FAIL_IF_NO_PSK, to be passed to
wolfSSL_CTX_set_verify(). When this option is enabled, a PSK is
mandatory to be negotiated between client and server, if the given
endpoint sets one.

This is relevant for external PSKs only (not session tickets) to make
sure the PSKs are used as an additional security factor (together with
ECDHE key exchange).

Also, when used with WOLFSSL_CERT_WITH_EXTERN_PSK, it makes sure that
peers are properly authenticated with both the PSK and via certificates.

Signed-off-by: Tobias Frauenschläger <tobias.frauenschlaeger@oth-regensburg.de>